Nodes Update
All nodes now have node type connections
- allows to signal node to node connection

Big redesign of nodes
- added interface
- allowed to be override

changed node registering to accommodate for the change
fixed bug in connection loading
cleaned up compiler warnings

new functionality isnt called or drawn yet tho

+++ src/raynodes/application/editor/EditorInit.h
@@ -50,7 +50,8 @@ inline void SetupDisplay(EditorContext& ec) {
   SetWindowMinSize(640, 480);
 
   // Center the window on the screen
-  SetWindowPosition((monitorWidth - screenWidth) / 2, (monitorHeight - screenHeight) / 2);
+  SetWindowPosition((static_cast<int>(monitorWidth) - screenWidth) / 2,
+                    (static_cast<int>(monitorHeight) - screenHeight) / 2);
 
   SetExitKey(0);
   DisableEventWaiting();
